[feature-20-login] remove clickhandling logic for nav pages from NavBar and move it to App
